Initial implementation. Has problem on x86-64 when compiled without -O2
Import files from discount-1.2.6. The library works fine compiled using the
rockspec on my 32 bit gcc 4.1.2 system. But when compiled on an x86-64 box
(gcc 4.3.1) without -O2, calling the function exported by ldiscount just
causes the program to acquire memory until it dies with a segmentation fault.
It appears to work fine when built with -O2 in the CFLAGS (as in the included
Makefile).