Fix infinite loop in populate function in mkdio.c
The problem can be summarised:
  /* defs */
  static char strget(void*); /* returns EOF, which is #defined as -1 */
  typedef unsigned int (*getc_func)(void*);

  /* function body extract */
  getc_func getc = (getc_func)strget;
  int c;
  while ( (c = (*getc)(ctx)) != EOF ) {...}

The cast of strget has undefined behaviour. On an x86-64 system with gcc
4.3.1, this code enters an infinite loop when compiled with -O0, as c is given
the value 255 and 255 != -1 is always true. This problem did not occur for -O1
and -O2. It also makes more sense for the typedef to be int. Fix by changing
the signature of strget to return int. It is also more sensible for the
getc_func typedef to be for return type int, but this commit does not make
that change.